# refactor(core): change impls of the `fetch_<CELLS>` methods