## Changes

Instead of iterating through darts, marking progressively each one, and adding new cell IDs to a list, we now compute cell IDs of all used darts. We can then collect the resulting iterator into a set, which automatically handles duplicates. 

This change effectively turns this operation from a search/coloring algorithm to a map_reduce operation.

## Why

The original implems could not be parallelized at all. They relied on sequential ordering, and used a single set to mark "visited" darts.

## Result

On a 512 by 512 square grid, the `fetch_<CELL>` method registered the following average execution time:

- before the optimization:
    - Vertices: 139.05 ms
    - Edges: 173.37 ms
    - Faces: 137.80 ms
- after the optimization:
    - Vertices: 98.795 ms
    - Edges: 65.829 ms
    - Faces: 67.801 ms

These respectively correspond to **29%, 62%, 51% decreases in execution time, at sequential execution**. The new implementation is also embarassingly parallel.