@imrn99 2024-10-22
> Is the set build also "embarrassingly parallel"?

I think I used "embarassingly parallel" loosely here. The set construction is most likely a reduction, which is a well defined parallel ops, though it doesn't scale ideally.